LayoutTests/media/track/track-webvtt-tc013-settings-expected.txt

Tests WebVTT settings.


*** Testing text track 0
EXPECTED (cues.length == '4') OK
EXPECTED (cues[i].linePosition == '100') OK
EXPECTED (cues[i].textPosition == '50') OK
EXPECTED (cues[i].alignment == 'middle') OK
EXPECTED (cues[i].direction == 'horizontal') OK
EXPECTED (cues[i].linePosition == '-1') OK
EXPECTED (cues[i].textPosition == '50') OK
EXPECTED (cues[i].alignment == 'middle') OK
EXPECTED (cues[i].direction == 'horizontal') OK
EXPECTED (cues[i].linePosition == '-1') OK
EXPECTED (cues[i].textPosition == '50') OK
EXPECTED (cues[i].alignment == 'middle') OK
EXPECTED (cues[i].direction == 'horizontal') OK
EXPECTED (cues[i].linePosition == '-1') OK
EXPECTED (cues[i].textPosition == '50') OK
EXPECTED (cues[i].alignment == 'middle') OK
EXPECTED (cues[i].direction == 'horizontal') OK

*** Testing text track 1
EXPECTED (cues.length == '4') OK
EXPECTED (cues[i].linePosition == '100') OK
EXPECTED (cues[i].textPosition == '50') OK
EXPECTED (cues[i].alignment == 'middle') OK
EXPECTED (cues[i].direction == 'horizontal') OK
EXPECTED (cues[i].linePosition == '-1') OK
EXPECTED (cues[i].textPosition == '50') OK
EXPECTED (cues[i].alignment == 'middle') OK
EXPECTED (cues[i].direction == 'horizontal') OK
EXPECTED (cues[i].linePosition == '-1') OK
EXPECTED (cues[i].textPosition == '50') OK
EXPECTED (cues[i].alignment == 'middle') OK
EXPECTED (cues[i].direction == 'horizontal') OK
EXPECTED (cues[i].linePosition == '-1') OK
EXPECTED (cues[i].textPosition == '50') OK
EXPECTED (cues[i].alignment == 'middle') OK
EXPECTED (cues[i].direction == 'horizontal') OK
END OF TEST
